+++ test/json-paths-utils.test.js
@@ -13,7 +13,16 @@ describe('JSONPaths Utilities', () => {
             const inputPath = path.resolve(__dirname, 'fixtures/object.json');
             const expectedOutputPath = path.resolve(__dirname, 'expected/paths.js');
             const input = JSON.parse(fs.readFileSync(inputPath, 'utf8'));
-            const output = getJSONPathsFromObject(input);
+            const output = await getJSONPathsFromObject(input);
+            const expectedOutput = (await import(expectedOutputPath)).paths;
+            expect(output).to.deep.equal(expectedOutput);
+        });
+
+        it('extracts the leaves JSONPaths from a JSON object', async () => {
+            const inputPath = path.resolve(__dirname, 'fixtures/object.json');
+            const expectedOutputPath = path.resolve(__dirname, 'expected/leaves.js');
+            const input = JSON.parse(fs.readFileSync(inputPath, 'utf8'));
+            const output = await getJSONPathsFromObject(input, true);
             const expectedOutput = (await import(expectedOutputPath)).paths;
             expect(output).to.deep.equal(expectedOutput);
         });
@@ -25,6 +34,15 @@ describe('JSONPaths Utilities', () => {
             const expectedOutputPath = path.resolve(__dirname, 'expected/paths.js');
             const input = JSON.parse(fs.readFileSync(inputPath, 'utf8'));
             const output = await getJSONPathsFromSchema(input);
+            const expectedOutput = (await import(expectedOutputPath)).paths;
+            expect(output).to.deep.equal(expectedOutput);
+        });
+
+        it('extracts the leaves JSONPaths from the properties of a JSON schema', async () => {
+            const inputPath = path.resolve(__dirname, 'fixtures/schema.json');
+            const expectedOutputPath = path.resolve(__dirname, 'expected/leaves.js');
+            const input = JSON.parse(fs.readFileSync(inputPath, 'utf8'));
+            const output = await getJSONPathsFromSchema(input, true);
             const expectedOutput = (await import(expectedOutputPath)).paths;
             expect(output).to.deep.equal(expectedOutput);
         });