+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+## [2.0.0] - 2024-11-24
+### Changed
+- Add boolean parameter `leaves` to `getJSONPathsFromObject` and `getJSONPathsFromSchema` functions.
+- The `getJSONPathsFromObject` function is now asynchronous.
+
 ## [1.0.1] - 2024-11-23
 ### Changed
 - Adjust inconsistent naming in mentions of `JSON Schema` and `JSONPath`.
@@ -10,6 +15,7 @@
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/alex-massa/extract-jsonpaths/compare/1.0.1...HEAD
+[Unreleased]: https://github.com/alex-massa/extract-jsonpaths/compare/2.0.0...HEAD
+[2.0.0]: https://github.com/alex-massa/extract-jsonpaths/compare/1.0.1...2.0.0
 [1.0.1]: https://github.com/alex-massa/extract-jsonpaths/compare/1.0.0...1.0.1
 [1.0.0]: https://github.com/alex-massa/extract-jsonpaths/commits/1.0.0